# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Krutrim Agent Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chats": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Create Chat",
        "operationId": "create_chat_api_chats_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "List Chats",
        "operationId": "list_chats_api_chats_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Chat"
                  },
                  "title": "Response List Chats Api Chats Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "Get Chat",
        "operationId": "get_chat_api_chats__chat_id__get",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "chats"
        ],
        "summary": "Update Chat",
        "operationId": "update_chat_api_chats__chat_id__put",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chats"
        ],
        "summary": "Delete Chat",
        "operationId": "delete_chat_api_chats__chat_id__delete",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDeletedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send Message",
        "operationId": "send_message_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendChatMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/agents/{agent_id}": {
      "post": {
        "summary": "Agent Run Endpoint",
        "operationId": "agent_run_endpoint_agents__agent_id__post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunAgentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityMessage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "const": "activity",
            "title": "Role",
            "default": "activity"
          },
          "activityType": {
            "type": "string",
            "title": "Activitytype"
          },
          "content": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "activityType",
          "content"
        ],
        "title": "ActivityMessage",
        "description": "An activity progress message emitted between chat messages."
      },
...
      "AssistantMessage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "const": "assistant",
            "title": "Role",
            "default": "assistant"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "encryptedValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encryptedvalue"
          },
          "toolCalls": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ToolCall"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toolcalls"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id"
        ],
        "title": "AssistantMessage",
        "description": "An assistant message."
      },
...
      "Chat": {
        "properties": {
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "sandbox_sharing": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "isolated",
                  "session-shared",
                  "project-shared"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Sharing"
          },
          "sandbox_idle_timeout_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Idle Timeout Seconds"
          },
          "sandbox_resource_overrides": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Resource Overrides"
          }
        },
        "type": "object",
        "required": [
          "chat_id",
          "display_name",
          "provider",
          "model",
          "created_at",
          "updated_at"
        ],
        "title": "Chat",
        "description": "A lightweight, non-agentic chat thread. `project_id` is optional; a standalone chat\nhas no meaningful sandbox policy until moved into a project."
      },
      "ChatApiMessage": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatApiMessage",
        "description": "One turn's worth of chat history \u2014 matches `shared-types.ts`'s\n`ChatApiMessage` and `chat/messages.py::from_lc_messages`'s output\nexactly. Used both by `POST /api/chat`'s reply and\n`GET /api/sessions/{id}/messages`'s history."
      },
      "ChatDeletedResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "chat_id"
        ],
        "title": "ChatDeletedResponse"
      },
      "ChatMessageRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "chat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "chat_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Title"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatMessageRequest"
      },
...
      "Context": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "description",
          "value"
        ],
        "title": "Context",
        "description": "Additional context for the agent."
      },
...
      "CreateChatRequest": {
        "properties": {
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "CreateChatRequest"
      },
...
      "DeveloperMessage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "const": "developer",
            "title": "Role",
            "default": "developer"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "encryptedValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encryptedvalue"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "content"
        ],
        "title": "DeveloperMessage",
        "description": "A developer message."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReasoningMessage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "const": "reasoning",
            "title": "Role",
            "default": "reasoning"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "encryptedValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encryptedvalue"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "content"
        ],
        "title": "ReasoningMessage",
        "description": "A reasoning message containing the agent's internal reasoning process."
      },
...
      "RunAgentInput": {
        "properties": {
          "threadId": {
            "type": "string",
            "title": "Threadid"
          },
          "runId": {
            "type": "string",
            "title": "Runid"
          },
          "parentRunId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentrunid"
          },
          "state": {
            "title": "State"
          },
          "messages": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/DeveloperMessage"
                },
                {
                  "$ref": "#/components/schemas/SystemMessage"
                },
                {
                  "$ref": "#/components/schemas/AssistantMessage"
                },
                {
                  "$ref": "#/components/schemas/UserMessage"
                },
                {
                  "$ref": "#/components/schemas/ToolMessage"
                },
                {
                  "$ref": "#/components/schemas/ActivityMessage"
                },
                {
                  "$ref": "#/components/schemas/ReasoningMessage"
                }
              ],
              "discriminator": {
                "propertyName": "role",
                "mapping": {
                  "activity": "#/components/schemas/ActivityMessage",
                  "assistant": "#/components/schemas/AssistantMessage",
                  "developer": "#/components/schemas/DeveloperMessage",
                  "reasoning": "#/components/schemas/ReasoningMessage",
                  "system": "#/components/schemas/SystemMessage",
                  "tool": "#/components/schemas/ToolMessage",
                  "user": "#/components/schemas/UserMessage"
                }
              }
            },
            "type": "array",
            "title": "Messages"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/Tool"
            },
            "type": "array",
            "title": "Tools"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/Context"
            },
            "type": "array",
            "title": "Context"
          },
          "forwardedProps": {
            "title": "Forwardedprops"
          },
          "resume": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ResumeEntry"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "threadId",
          "runId",
          "state",
          "messages",
          "tools",
          "context",
          "forwardedProps"
        ],
        "title": "RunAgentInput",
        "description": "Input for running an agent."
      },
      "SendChatMessageResponse": {
        "properties": {
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "message": {
            "$ref": "#/components/schemas/ChatApiMessage"
          }
        },
        "type": "object",
        "required": [
          "chat_id",
          "session_id",
          "message"
        ],
        "title": "SendChatMessageResponse"
      },
...
      "SystemMessage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "const": "system",
            "title": "Role",
            "default": "system"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "encryptedValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encryptedvalue"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "content"
        ],
        "title": "SystemMessage",
        "description": "A system message."
      },
...
      "Tool": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "parameters": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Parameters"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "name",
          "description"
        ],
        "title": "Tool",
        "description": "A tool definition."
      },
...
      "ToolMessage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "const": "tool",
            "title": "Role",
            "default": "tool"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "toolCallId": {
            "type": "string",
            "title": "Toolcallid"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "encryptedValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encryptedvalue"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "content",
          "toolCallId"
        ],
        "title": "ToolMessage",
        "description": "A tool result message."
      },
...
      "UpdateChatRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "type": "object",
        "title": "UpdateChatRequest"
      },
...
      "UserMessage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "const": "user",
            "title": "Role",
            "default": "user"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/TextInputContent"
                    },
                    {
                      "$ref": "#/components/schemas/ImageInputContent"
                    },
                    {
                      "$ref": "#/components/schemas/AudioInputContent"
                    },
                    {
                      "$ref": "#/components/schemas/VideoInputContent"
                    },
                    {
                      "$ref": "#/components/schemas/DocumentInputContent"
                    },
                    {
                      "$ref": "#/components/schemas/BinaryInputContent"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "audio": "#/components/schemas/AudioInputContent",
                      "binary": "#/components/schemas/BinaryInputContent",
                      "document": "#/components/schemas/DocumentInputContent",
                      "image": "#/components/schemas/ImageInputContent",
                      "text": "#/components/schemas/TextInputContent",
                      "video": "#/components/schemas/VideoInputContent"
                    }
                  }
                },
                "type": "array"
              }
            ],
            "title": "Content"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "encryptedValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encryptedvalue"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "content"
        ],
        "title": "UserMessage",
        "description": "A user message supporting text or multimodal content."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}